+++ src/resources/tasks/task.memory.repository.ts
@@ -1,5 +1,5 @@
 import { tasksDB } from '../../common/memoDB';
-import { ITask } from '../../common/types';
+import { ITask } from '../../common/interfaces';
 
 /**
  * Returns the list of created tasks
